refactor: move PulseInsightsApiError to separate file per review feedback
